test: Add instrumented test for DatabaseService.clearAll
Adds an instrumented test that inserts data into a real Realm database
and verifies it is correctly deleted by DatabaseService.clearAll().
This replaces the earlier mock-interaction test that was vacuous
due to the inability to initialize Realm in a pure JVM unit test.

Also fixes a compilation error in RealmUserTest.kt by injecting
the missing Lazy<ResourcesRepository> and Lazy<CoursesRepository>
dependencies into UserRepositoryImpl.
